+++ src/modules/user/repositories/user.repository.ts
@@ -1,10 +1,98 @@
 import { Injectable } from '@nestjs/common';
-import { ActiveStatus, Sex } from '@prisma/client';
+import {
+  ActiveStatus,
+  AddressLevel,
+  AuthProvider,
+  Prisma,
+  Sex,
+} from '@prisma/client';
 import { PrismaService } from '../../../infra/prisma/prisma.service';
 
 @Injectable()
 export class UserRepository {
   constructor(private readonly prismaService: PrismaService) {}
+
+  async upsertKakaoUser({
+    providerUserId,
+    email,
+    nickname,
+    defaultBirthdate,
+    defaultAddressCode,
+    defaultIntroVoiceUrl,
+    defaultProfileImageUrl,
+  }: {
+    providerUserId: string;
+    email: string;
+    nickname: string;
+    defaultBirthdate: Date;
+    defaultAddressCode: string;
+    defaultIntroVoiceUrl: string;
+    defaultProfileImageUrl: string;
+  }) {
+    await this.ensureDefaultAddress(defaultAddressCode);
+
+    const where = {
+      provider_providerUserId: {
+        provider: AuthProvider.KAKAO,
+        providerUserId,
+      },
+    };
+
+    const existing = await this.prismaService.user.findUnique({ where });
+    if (existing) {
+      const updatedUser = await this.prismaService.user.update({
+        where,
+        data: {
+          email,
+        },
+      });
+
+      return { user: updatedUser, isNewUser: false };
+    }
+
+    const insertedRows = await this.prismaService.$queryRaw<
+      Array<{ id: bigint }>
+    >(Prisma.sql`
+      INSERT INTO "User" (
+        "birthdate",
+        "email",
+        "nickname",
+        "updatedAt",
+        "introVoiceUrl",
+        "introText",
+        "profileImageUrl",
+        "code",
+        "provider",
+        "providerUserId",
+        "vibeVector"
+      )
+      VALUES (
+        ${defaultBirthdate},
+        ${email},
+        ${nickname},
+        NOW(),
+        ${defaultIntroVoiceUrl},
+        ${''},
+        ${defaultProfileImageUrl},
+        ${defaultAddressCode},
+        ${AuthProvider.KAKAO}::"AuthProvider",
+        ${providerUserId},
+        '[0]'::vector
+      )
+      RETURNING "id"
+    `);
+
+    const createdId = insertedRows[0]?.id;
+    if (!createdId) {
+      return null;
+    }
+
+    const createdUser = await this.prismaService.user.findUniqueOrThrow({
+      where: { id: createdId },
+    });
+
+    return { user: createdUser, isNewUser: true };
+  }
 
   findProfileById(userId: number) {
     return this.prismaService.user.findFirst({
@@ -169,6 +257,27 @@ export class UserRepository {
       data: {
         status: ActiveStatus.INACTIVE,
         deletedAt: new Date(),
+      },
+    });
+  }
+
+  private async ensureDefaultAddress(defaultAddressCode: string) {
+    await this.prismaService.address.upsert({
+      where: { code: defaultAddressCode },
+      update: {},
+      create: {
+        code: defaultAddressCode,
+        sidoCode: '00',
+        sigunguCode: '000',
+        emdCode: '000',
+        riCode: '00',
+        fullName: 'Unknown',
+        sidoName: 'Unknown',
+        sigunguName: null,
+        emdName: null,
+        riName: null,
+        level: AddressLevel.SIGUNGU,
+        parentCode: null,
       },
     });
   }
